relocate files in /app and create /stores
