+++ src/components/PostContainer.tsx
@@ -18,8 +18,8 @@ import { ellipsisVertical, share, trash } from 'ionicons/icons';
 import React, { ReactElement, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useHistory } from 'react-router';
-import { toast } from '../app/toast';
-import { selectUser } from '../features/User/UserStore';
+import { toast } from './data/toast';
+import { selectUser } from '../stores/users/user.store';
 import { IPost } from '../services/models/DocTypes.model';
 import { db, Timestamp } from '../services/data/firebase';
 import { HomePost } from './HomePostView';
