+++ src/components/PrivateRoute.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Login from '../pages/LoginPage';
 import { useSelector } from 'react-redux';
-import { selectUser } from '../features/User/UserStore';
+import { selectUser } from '../stores/users/user.store';
 
 type PrivateRouteTypes = {
     component: React.FC;
