--- src/features/counter/Counter.scss
@@ -1,9 +1,0 @@
-.counterList {
-    width: fit-content;
-    margin: auto;
-}
-
-.counterText {
-    align-content: space-between;
-    margin: auto;
-}
