+++ src/pages/LoginPage.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { InputChangeEventDetail } from '@ionic/core';
 import { IonContent, IonPage, IonInput, IonButton, IonRouterLink } from '@ionic/react';
 import '../styles/pages/Login.scss';
-import { toast } from '../app/toast';
-import { FirebaseAuthService } from '../services/FirebaseAuth.service';
+import { toast } from '../components/data/toast';
+import FirebaseAuthService from '../services/FirebaseAuth.service';
 import Footer from '../components/Footer';
 
 /**
