+++ src/stores/users/user.store.ts
@@ -1,9 +1,9 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-// import { AppThunk, RootState } from '../../app/store';
-import User from '../../services/User.service';
+import PrimaryUser from '../../services/PrimaryUser.service';
+import { RootState } from '../store';
 
 interface UserState {
-    user: User | undefined;
+    user: PrimaryUser | undefined;
 }
 
 const initialState: UserState = {
@@ -14,7 +14,7 @@ export const slice = createSlice({
     name: 'user',
     initialState,
     reducers: {
-        setUser: (state, action: PayloadAction<User>) => {
+        setUser: (state, action: PayloadAction<PrimaryUser>) => {
             state.user = action.payload;
         },
         clearUser: (state) => {
@@ -38,6 +38,6 @@ export const { setUser, clearUser } = slice.actions;
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
-// export const selectCount = (state: RootState): number => state.counter.value;
+export const selectUser = (state: RootState): PrimaryUser | undefined => state.user.user;
 
 export default slice.reducer;
